+++ feature_layers/configure-clusters/README.metadata.json
@@ -21,8 +21,7 @@
       "FeatureLayer",
       "FeatureReduction",
       "GeoElement",
-      "IdentifyLayerResult",
-      "PopupViewer"
+      "IdentifyLayerResult"
     ],
     "redirect_from": [
         "/java/latest/sample-code/add-feature-clustering.htm"
@@ -33,8 +32,7 @@
       "FeatureLayer",
       "FeatureReduction",
       "GeoElement",
-      "IdentifyLayerResult",
-      "PopupViewer"
+      "IdentifyLayerResult"
     ],
     "snippets": [
         "src/main/java/com/esri/samples/add_feature_clustering/AddFeatureClusteringSample.java"
